Added heap state structs to allocator fuzz test
This introduces a new layer of abstraction in the allocator test
to make it easier to fuzz different allocator capabilities. The new
class associates a heap capability with cached free vectors.

The extra layer of abstraction lets us add additional heap caps to
the fuzz test without much manual effort. This also enables fuzzing
of heap_free_all, since we cannot invoke that on the default allocator
capability without freeing the memory we're using for test state.