Added fuzzing for `heap_free_all`
Implemented a `doFreeAll` helper in the allocator fuzz test case that
caches a number of frees (much like `doFree`) before freeing all.
This function is called ~10% of the time in place of the prior random number
of `doFree` calls, with the intention being that the functionlity is
exercised without either massively increasing the runtime of the fuzz
test case or detracting from the fuzzing of `heap_free` calls.

We do not invoke `heap_free` on MALLOC_CAPABILITY as that causes
invariant failures (EXPAND ON THIS -- WHY?)

Further work can be done testing how this interacts with sealed
allocations, which won't be freed by `heap_free_all`.