Updated `heap_can_free` documentation to reflect return values
The comments now reflect the possible error values it can return.
Specifically, we return `-EPERM` if `heapCapability` is not valid
or it cannot free the provided allocation. This can happen when there
are existing claims on the quota when an owner tries to free.
`-EINVAL` is returned if `ptr` is not a valid pointer into a live
heap allocation.