Added allocator permissions on top of handle bit encoding
This adds support for using the permissions encoded in the low
2 bits of the allocator capability to confer allocation/freeing
permission, supported with checks throughout the allocator APIs.
Scheme is as follows:

|bit 1|bit 2| Permissions        |
|-----|-----|--------------------|
|  1  |  1  | fully permissioned |
|  1  |  0  | can free, no alloc |
|  0  |  1  | can alloc, no free |
|  0  |  0  | no alloc or free   |

If an allocator capability does not carry with it permission to
allocate, it also cannot be used to claim. Similarly, if it cannot be
used to free, it cannot be used to release a claim.

Permissions are extracted during `malloc_capability_unseal`, since the
process of unsealing clobbers their values. We return permissions as the
second element in a std::pair from this function.

Notably, `heap_quota_remaining` is permission agnostic.
`allocate_sealed_unsealed` for now only checks `allocate` permissions,
but could be extended to use the third available bit in the handle to
further restrict the ability of a quota to be used to make sealed
allocations.